+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Google Calendar Change Log
 ==========================
 
  * Ported to Chromium backend.
+ * If desktop notifications are enabled, they are used instead of web app alerts.
 
 1.2 - February 12th, 2017
 -------------------------